+++ bin/src/utils/config_loader.dart
@@ -7,7 +7,7 @@ class ConfigLoader {
     final value = getValue(doc, key);
     if (value != null) {
       if (value is String) return value;
-      CliLog.stepWarn("'$key' is invalid");
+      CliLog.warn("'$key' is invalid");
     }
     return defaultValue;
   }
@@ -20,7 +20,7 @@ class ConfigLoader {
           if (item is String) {
             set.add(item);
           } else {
-            CliLog.stepWarn("'$key' has invalid item '$value'. Must be of type String.");
+            CliLog.warn("'$key' has invalid item '$value'. Must be of type String.");
           }
         }
       } else if (value is List<String>) {
@@ -28,7 +28,7 @@ class ConfigLoader {
       } else if (value is String) {
         set.add(value);
       } else {
-        CliLog.stepWarn("'$key' is of type ${value.runtimeType}. Must be of type YamlList, List<String>, or String.");
+        CliLog.warn("'$key' is of type ${value.runtimeType}. Must be of type YamlList, List<String>, or String.");
       }
     }
   }
@@ -41,13 +41,13 @@ class ConfigLoader {
           if (entry.value is String) {
             map[entry.key] = entry.value;
           } else {
-            CliLog.stepWarn("'$key' has invalid item '$value'");
+            CliLog.warn("'$key' has invalid item '$value'");
           }
         }
       } else if (value is Map<String, String>) {
         map.addAll(value);
       } else {
-        CliLog.stepWarn("'$key' is of type ${value.runtimeType}. Must be of type YamlMap or Map<String, String>.");
+        CliLog.warn("'$key' is of type ${value.runtimeType}. Must be of type YamlMap or Map<String, String>.");
       }
     }
   }