+++ bin/src/utils/utils.dart
@@ -1,0 +1,7 @@
+bool isBlank(String? value) {
+  return value == null || value.isEmpty;
+}
+
+bool isNotBlank(String? value) {
+  return !isBlank(value);
+}