+++ test/fixtures/inflater_spec.dart
@@ -1,0 +1,14 @@
+import 'package:xwidget/src/custom/async.dart';
+import 'package:xwidget/src/custom/controller.dart';
+import 'package:xwidget/src/custom/list_of.dart';
+import 'package:xwidget/src/custom/value_listener.dart';
+
+// ignore_for_file: unused_element
+
+// Define constants for all widgets and objects we want to generate inflaters for.
+// Used by the builder to generate inflaters.
+
+const ControllerWidget? _controllerWidget = null;
+const DynamicBuilder? _dynamicBuilder = null;
+const ListOf? _listOf = null;
+const ValueListener? _valueListener = null;