+++ test/lib/utils/brackets_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:xwidget/utils/brackets.dart';
+
+import 'package:xwidget/xwidget.dart';
 
 
 main() {