+++ test/lib/utils/data_test.dart
@@ -1,5 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:xwidget/utils/brackets.dart';
+
+import 'package:xwidget/xwidget.dart';
 
 
 main() {