+++ test/lib/xwidget_test.dart
@@ -1,9 +1,10 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:xwidget/utils/parsers.dart';
+
+import 'package:xwidget/src/utils/parsers.dart';
 
 
 void main() {
-  test('adds one to input values', () {
+  test('test bool parse', () {
     parseBool("true");
   });
 }